refactor: manager 뮤테이션 훅 구조 정리
